requirements.txt
mcp>=1.3.0
kiteconnect>=4.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
nsepython>=2.8
yfinance>=0.2.50
feedparser>=6.0.0
gnews>=0.4.0
